lima: Update to 0.14.2
Changes:
0.14.2
------
Bugfix release.

0.14.1
------
This release fixes a regression #1242 in Lima v0.14.0:  On colima,
docker data, such as containers and images in /var/lib/docker were
unmounted and they looked as if they were removed.

The actual data are not removed, and can be rescued by upgrading Lima
to this v0.14.1 release (or downgrading to v0.13.0).

Only the following instances are known to have been affected:

- colima instances
- alpine instances that were created with Lima before v0.13.0

The default (ubuntu) instances are unaffected.

0.14.0
------
This release adds an experimental support for Apple's
Virtualization.framework aka vz.