helm: Update to 3.11.1
Changes:
3.11.1
------
Helm v3.11.1 is a security (patch) release. Users are strongly
recommended to update to this release.

The template function `getHostByName` can be used to disclose
information. More details are available in the
CVE <https://github.com/helm/helm/security/advisories/GHSA-pwcw-6f5g-gxf8>.

This release introduces a breaking changes to Helm:

* When using the `helm` client for the `template`, `install`, and
  `upgrade` commands there is a new flag. `--enable-dns` needs to be set
  for the `getHostByName` template function to attempt to lookup an IP
  address for a given hostname. If the flag is not set the template
  function will return an empty string and skip looping up an IP address
  for the host.
* The Helm SDK has added the `EnableDNS` property to the install
  action, the upgrade action, and the `Engine`. This property must be set
  to true for the in order for the `getHostByName` template function to
  attempt to lookup an IP address.

The default for both of these cases is false.

Philipp Stehle at SAP disclosed the vulnerability to the Helm project.

3.11.0
------
Helm v3.11.0 is a feature release. Users are encouraged to upgrade for
the best experience.

- The Helm status command and the SDK can now show the status of core
  resources deployed in a chart (e.g., deployments). To use with `helm
  status` you need to use the `--show-resources` flag.
- Add support for comma separated values in template --api-versions
- Allow CGO_ENABLED to be overridden when building Helm from source