chirp: Add version 20230122
This is "chirp next" which is the gtk3 port.  It is command-line only
in pkgsrc because wxWidgets is out of date.

CHIRP is a free, open-source tool for programming your amateur radio.
It supports a large number of manufacturers and models, as well as
provides a way to interface with multiple data sources and formats.

This package provides the stable version preferred by upstream,
sometimes called "chirp-next" in other contexts.