perl: update to 5.37.9.
= Core Enhancements

== New C<class> Feature

A new B<experimental> syntax is now available for defining object classes,
where per-instance data is stored in "field" variables that behave like
lexicals.

    use feature 'class';

    class Point
    {
        field $x;
        field $y;

        method zero { $x = $y = 0; }
    }

This is described in more detail in L<perlclass>.  Notes on the internals of
its implementation and other related details can be found in L<perlclassguts>.

This remains a new and experimental feature, and is very much still under
development. It will be the subject of much further addition, refinement and
alteration in future releases.  As it is experimental, it yields warnings in
the C<experimental::class> category.  These can be silenced by a
C<no warnings> statement.

    use feature 'class';
    no warnings 'experimental::class';

== REG_INF has been raised from 65,536 to 2,147,483,647

Many regex quantifiers used to be limited to U16_MAX in the past, but are
now limited to I32_MAX, thus it is now possible to write /(?:word){1000000}/
for example.  Note that doing so may cause the regex engine to run longer
and use more memory.

== New API functions optimize_optree and finalize_optree

There are two new API functions for operating on optree fragments, ensuring
you can invoke the required parts of the optree-generation process that might
otherwise not get invoked (e.g. when creating a custom LOGOP).  To get access
to these functions, you first need to set a C<#define> to opt-in to using
these functions.

  #define PERL_USE_VOLATILE_API

These functions are closely tied to the internals of how the interpreter
works, and could be altered or removed at any time if other internal changes
make that necessary.

= Incompatible Changes

== (**{ ... }) removed from the regex engine.

This feature was released as part of 5.37.8, after some use and
discussion it was seen as more problematic than understood at first
and has been removed in 5.37.9. It was only ever present in a single
development release and has never been released as part of a production perl,
thus no deprecation cycle has been performed.

= Deprecations

== Use of C<'> as a package name separator is deprecated

Using C<'> as package separator in a variable named in a double-quoted
string has warned since 5.28.  It is now deprecated in both string
interpolation and non-interpolated contexts, and will be removed in
Perl 5.40.

Also Performance Enhancements, and updates to Modules and Pragmata.