Import libxo 1.6.0 as wip/libxo
The libxo library allows an application to generate text, XML, JSON, and
HTML output using a common set of function calls. The application decides at
run time which output style should be produced. The application calls a
function "xo_emit" to product output that is described in a format string. A
"field descriptor" tells libxo what the field is and what it means.