kubectl: Update to 1.26.2
API Change

* A fix in the resource.k8s.io/v1alpha1/ResourceClaim API avoids
harmless (?) ".status.reservedFor: element 0: associative list without keys
has an element that's a map type" errors in the apiserver. Validation now
rejects the incorrect reuse of the same UID in different entries.
* K8s.io/component-base/logs: usage of the pflag values in a normal Go flag
set led to panics when printing the help message

Feature

* Kubelet TCP and HTTP probes are more effective using networking
resources: conntrack entries, sockets, ... This is achieved by reducing the
TIME-WAIT state of the connection to 1 second, instead of the defaults 60
seconds. This allows kubelet to free the socket, and free conntrack entry
and ephemeral port associated.
* Kubernetes is now built with Go 1.19.6
* Use HorizontalPodAutoscaler v2 for kubectl

Bug or Regression

* Do not add DisruptionTarget condition by PodGC for pods which are in
terminal phase
* Enforce nodeName cannot be set along with non-empty
schedulingGates
* Fix a bug that caused to panic the apiserver when trying to allocate a
Service with a dynamic ClusterIP and it has been configured with Service
CIDRs with a /28 mask for IPv4 and a /124 mask for IPv6
* Fix nil pointer error in nodevolumelimits csi logging
* Fix the regression that introduced 34s timeout for DELETECOLLECTION
calls
* Fixed bug which caused the status of Indexed Jobs to only be updated when
there are newly completed indexes. The completed indexes are now updated if
the .status.completedIndexes has values outside of the [0,
.spec.completions> range
* Fixes bug in ValidatingAdmissionPolicy alpha which prevented policies
from using a paramKind previously used by another policy
* Golang.org/x/net updates to v0.7.0 to fix CVE-2022-41723
* The Kubernetes API server now correctly detects and closes existing TLS
connections when its client certificate file for kubelet authentication has
been rotated.