kubectl: Update to 1.26.3
API Change

* Volumes: resource.claims gets cleared for PVC specs during create or
update of a pod spec with inline PVC template or of a PVC because it has no
effect.

Feature

* Kubernetes is now built with Go 1.19.7
* The go version defined in .go-version is now fetched when invoking test,
build, and code generation targets if the current go version does not match
it. Set $FORCE_HOST_GO=y while testing or building to skip this behavior,
or set $GO_VERSION to override the selected go version.

Failing Test

* Fixed panic in vSphere e2e tests.

Bug or Regression

* Fix data race in kube-scheduler when preemption races with a Pod
update.
* Fix log line in scheduler that inaccurately implies that volume binding
has finalized
* Fix race in alpha aggregated discovery handler Yes, discovery document
will correctly return the resources for aggregated apiservers that do not
implement aggregated disovery
* Fixed a bug where Kubernetes would apply a default StorageClass to a
PersistentVolumeClaim, even when the deprecated annotation
volume.beta.kubernetes.io/storage-class was set.
* Fixed an EndpointSlice Controller hashing bug that could cause
EndpointSlices to incorrectly handle Pods with duplicate IP addresses. For
example this could happen when a new Pod reused an IP that was also
assigned to a Pod in a completed state.
* Fixed performance regression in scheduler caused by frequent metric
lookup on critical code path.
* Fixing issue with Winkernel Proxier - ClusterIP Loadbalancers are missing
if the ExternalTrafficPolicy is set to Local and the available endpoints
are all remoteEndpoints.
* Fixing issue with Winkernel Proxier - IPV6 load balancer policies are
missing when service is configured with ipFamilyPolicy:
RequireDualStack
* Make kubectl diff --prune behave correctly with the --selector/-l
flag
* Remove check for CSI driver running on node for CSI migration
* Set device stage path whenever available for expansion during mount