mrust: new sub-package "mrust-rustclibs"
mrust-rustclibs provides the standard rust libraries needed
by mrust-rustc to compile a .rs file (e.g.,
rustc-1.54.0-src/src/test/ui/hello.rs) into a working
executable.  The build of the pkg libraries is coordinated
by minicargo (as opposed to real cargo), but the actual code
is compiled by mrust-rustc (as opposed to mrustc).

Note that mrust-rustc, approximately, IS rustc: it's a full
build of, currently, version 1.54.0 of the compiler--from an
almost pristine rust-lang.org src tree (although the build
is of course done with mrustc & minicargo rather than with
an existing, official rustc & cargo, and there are therefore
a handful of needed patches applied to the tree first... see
rustc-1.54.0-src.patch in the mrustc src dist).  mrustc,
OTOH, is a stripped-down, C++ implementation of a rust
compiler, developed mainly for bootstrapping rustc from
(rust-lang.org) src without rust binaries.

My hope, as a result, is that between mrust-rustc,
mrust-cargo & mrust-rustclibs, there is enough scaffolding
to bootstrap to rust 1.55 via the rust-lang.org srcs & the
appropriate vintage of pkgsrc/lang/rust; except, in this
case, without ANY other rust infrastructure: i.e., no
rust-lang.org binaries, nor a NetBSD rustc binary bootstrap
kit.

Wisdom from the author of mrustc, however, suggests that a
few more stages of the mrustc -> ... -> rustc 1.54 bootstrap
will likely be needed--so as to include some final elements
of a fully functional rustc toolchain--before the above will
work.  If that turns out to be the case, mrust-rustclibs
will need to be updated to do a bunch more work...