consul: Update to 1.15.2
FEATURES:

* xds: Allow for configuring connect proxies to send service mesh telemetry
to an HCP metrics collection service.

BUG FIXES:

* audit-logging: (Enterprise only) Fix a bug where /agent/monitor and
/agent/metrics endpoints return a Streaming not supported error when audit
logs are enabled. This also fixes the delay receiving logs when running
consul monitor against an agent with audit logs enabled.
* ca: Fixes a bug where updating Vault CA Provider config would cause TLS
issues in the service mesh
* cache: revert cache refactor which could cause blocking queries to never
return
* gateway: (Enterprise only) Fix bug where namespace/partition would fail
to unmarshal for TCPServices.
* gateway: (Enterprise only) Fix bug where namespace/partition would fail
to unmarshal.
* gateway: (Enterprise only) Fix bug where parent refs and service refs for
a route in the same namespace as the route would fallback to the default
namespace if the namespace was not specified in the configuration rather
than falling back to the routes namespace.
* gateway: (Enterprise only) Fix bug where routes defined in a different
namespace than a gateway would fail to register.
* gateways: Adds validation to ensure the API Gateway has a listener
defined when created
* gateways: Fixes a bug API gateways using HTTP listeners were taking
upwards of 15 seconds to get configured over xDS.
* peering: (Consul Enterprise only) Fix issue where connect-enabled
services with peer upstreams incorrectly required service:write access in
the default namespace to query data, which was too restrictive. Now having
service:write to any namespace is sufficient to query the peering data.
* peering: (Consul Enterprise only) Fix issue where resolvers, routers, and
splitters referencing peer targets may not work correctly for non-default
partitions and namespaces. Enterprise customers leveraging peering are
encouraged to upgrade both servers and agents to avoid this problem.
* peering: Fix issue resulting in prepared query failover to cluster peers
never un-failing over.
* peering: Fixes a bug that can lead to peering service deletes impacting
the state of local services
* peering: Fixes a bug where the importing partition was not added to
peered failover targets, which causes issues when the importing partition
is a non-default partition.
* raft_logstore: Fixes a bug where restoring a snapshot when using the
experimental WAL storage backend causes a panic.
* ui: fix PUT token request with adding missed AccessorID property to
requestBody
* ui: fix rendering issues on Overview and empty-states by addressing
isHTMLSafe errors