zoneminder: Patch for NetBSD/earmv7hf-el
Patch more types, mostly time_t which is not long.

Do not assume linux if arm.  Rather than using a linux syscall to
determine if neon is present, just assume it is, optimizing for RPI2/3
and hoping little harm is done, as zm on a RPI1 is a stretch anyway
due to lack of RAM.