regal: Import regal-0.2.0 as wip/regal
Regal is a linter for Rego, with the goal of making your Rego
magnificent!

Regal goals are:

- Identify common mistakes, bugs and inefficiencies in Rego policies,
  and suggest better approaches
- Provide advice on best practices, coding style, and tooling
- Allow users, teams and organizations to enforce custom rules on their
  policy code

Regal rules are to as large extent as possible written in Rego
themselves, using the JSON representation of the Rego abstract syntax
tree (AST) as input, a few additional custom built-in functions and
some indexed data structures to help with linting.