py-wxWidgets: Start to update
This is hacky; just enough to communicate to others where I think the
new upstream is, for the software that I think will be accepted by
things that currently use old py-wxWidgets.  This is at the moment
theory.

It appears that it just uses gtk, and there is no need for wxGTK.