prometheus-lts: Update to 2.45.0
pkgsrc changes:
- Switch to use upstream "web-ui" distfile so that we avoid building and
  uploading the React UI at every update
- Some possibly improvements / lint still needed but probably near to be
  okay-ish to import it in pkgsrc too

Changes:
2.45.0 / 2023-06-23
This release is a LTS (Long-Term Support) release of Prometheus and will
receive security, documentation and bugfix patches for at least 12 months.
Please read more about our LTS release cycle at
<https://prometheus.io/docs/introduction/release-cycle/>.

2.44.0 / 2023-05-13
This version is built with Go tag `stringlabels`, to use the smaller data
structure for Labels that was optional in the previous release.

2.43.0 / 2023-03-21
We are working on some performance improvements in Prometheus, which are only
built into Prometheus when compiling it using the Go tag `stringlabels`
(therefore they are not shipped in the default binaries). It uses a data
structure for labels that uses a single string to hold all the label/values,
resulting in a smaller heap size and some speedups in most cases. We would like
to encourage users who are interested in these improvements to help us measure
the gains on their production architecture. We are providing release artefacts
`2.43.0+stringlabels` and Docker images tagged `v2.43.0-stringlabels` with those
improvements for testing.

2.42.0 / 2023-01-31
This release comes with a bunch of feature coverage for native
histograms and breaking changes.

If you are trying native histograms already, we recommend you remove
the `wal` directory when upgrading.  Because the old WAL record for
native histograms is not backward compatible in v2.42.0, this will lead
to some data loss for the latest data.

Additionally, if you scrape "float histograms" or use recording rules
on native histograms in v2.42.0 (which writes float histograms), it is
a one-way street since older versions do not support float histograms.

2.40.0 / 2022-11-08
This release introduces an experimental, native way of representing and
storing histograms.

It can be enabled in Prometheus via
`--enable-feature=native-histograms` to accept native histograms.
Enabling native histograms will also switch the preferred exposition
format to protobuf.

To instrument your application with native histograms, use the v1.14.0
or later of `client_golang` and set the `NativeHistogramBucketFactor`
in your `HistogramOpts` (`1.1` is a good starting point).  Your
existing histograms won't switch to native histograms until
`NativeHistogramBucketFactor` is set.