mpc: Import mpc-1.3.1 as wip/mpc
MPC is a complex floating-point library with exact rounding.
It is based on the GNU MPFR floating-point library, which is itself
based on the GNU MP library.

This is a C library for the arithmetic of complex numbers with
arbitrarily high precision and correct rounding of the result.
It extends the principles of the IEEE-754 standard for fixed
precision real floating point numbers to complex numbers,
providing well-defined semantics for every operation.
At the same time, speed of operation at high precision is a
major design goal.