kubepug: Update to 1.5.0
Changes:
v1.5.0
------
This is Kubepug v1.5.0!!

It was mostly rewritten for performance and better results! Thanks for
using the project!!

More great things are coming for next version!

The api-walk flag was removed! There is no need for the flag anymore,
and Kubepug will always validate against all resources!

The main logic of program was simplified and rewritten for better
performance. A lot of false positives have been removed! A bunch of
bugs squashed!

This release was also a preparation for a bigger one, that will allow
better tuning of APIs to be validated, and a much reliable API
database. Also, it allows a better extension, so someone willing to use
Kubepug as a library, a webhook, or even writing their own output
plugin will be able to do it!