rust: update to version 1.71.0.
This is slightly unfinished...  A native build on NetBSD/amd64 succeeds,
but a cross-build (anything which builds the internal LLVM) will fail
due to RPATH issues with args containing $ORIGIN/../lib.  See TODO for
details.

Once that issue is sorted out, there may be further checksum substitutions
which need updating in Makefile.