Import Hercules 2023.07.12 as wip/hercules-server
Hercules is a collaborative software development project revolving around
the creation of a robust Massively Multiplayer Online Role-Playing Game
(MMORPG) server package. Written in C, the program is very versatile and
provides NPCs, warps and modifications. The project is jointly managed by a
group of volunteers located around the world as well as a tremendous
community providing QA and support. Hercules is a continuation of the
original Athena project.