hercules4aethra: add new package
For now this is preliminary since there are no releases yet.
Also because it starts out as a straight copy of SDL Hercules, I copied
the packaging for emulators/hercules4sdl. Because of this simple
approach, both packages install exactly the same file names.
So they conflict massively. Maybe this can be improved later on.