regal: Update to 0.6.0
Changes:
0.6.0
-----
This release brings a new command for quickly generating new (custom or
built-in) rules, a new linter rule, and some improvements around
tooling.

0.5.0
-----
This release brings improvements and new features to improve the
experience of authoring custom rules, as well as new, granular
capabilities for ignoring files. Most of these improvements are
directly based on feedback — and in some cases contributions — from
Regal users, which is particularly exciting!

0.4.0
-----
This release brings three new rules related to comments and metadata annotations:

- invalid-metadata-attribute
- detached-metadata
- no-whitespace-comment

Additionally, new end-to-end tests exposed a few mistakes in a previous
refactoring, which have been fixed. This mistake meant that v0.3.0
failed to correctly run the `line-length` and `function-arg-return`
rules... so if you started from that release you're really getting
*five* new rules with v0.4.0... good thing we're keeping a fast paced
release cadence! Thanks to @kristiansvalland for reporting on this
regression.