Import SDL2_testgamecontroller
This package provides the game controller test applications from SDL. You
can plug in a joystick or gamepad to validate its functionality. The
programs also will generate a controller database mapping string for you,
which is usually necessary in order to get SDL-based apps to recognize your
device.

(In reality this probably should be an option of devel/SDL2 but I'm still
experimenting with hardware at the moment)