u-boot-tools: fix and update to v2023.07.
The previous version of this package did not build correctly, because
it would not link against the SSL library.  Add the buildlink3.mk file
and -L to the definition of HOSTLDLIBS.  While at it, update to
v2023.07 and clean up the package (e.g., align with the contents of
sysutils/u-boot/u-boot.mk, add man pages, add a test).