opentofu: Import opentofu-1.6.0-alpha2 as wip/opentofu
OpenTofu is an OSS tool for building, changing, and versioning
infrastructure safely and efficiently. OpenTofu can manage existing and
popular service providers as well as custom in-house solutions.

The key features of OpenTofu are:

* Infrastructure as Code: Infrastructure is described using a
  high-level configuration syntax. This allows a blueprint of your
  datacenter to be versioned and treated as you would any other code.
  Additionally, infrastructure can be shared and re-used.
* Execution Plans: OpenTofu has a "planning" step where it generates
  an execution plan. The execution plan shows what OpenTofu will do
  when you call apply. This lets you avoid any surprises when
  OpenTofu manipulates infrastructure.
* Resource Graph: OpenTofu builds a graph of all your resources, and
  parallelizes the creation and modification of any non-dependent
  resources. Because of this, OpenTofu builds infrastructure as
  efficiently as possible, and operators get insight into dependencies in
  their infrastructure.
* Change Automation: Complex changesets can be applied to your
  infrastructure with minimal human interaction. With the previously
  mentioned execution plan and resource graph, you know exactly what
  OpenTofu will change and in what order, avoiding many possible human
  errors.

pkgsrc terraform bits were adjusted in a very naif way and some
FIXME/XXX were left! Everything still not tested, beware!

(But this should be hopefully a good starting point in order to have a
generic Terraform agnostic pkgsrc mk infrastructure and then permit
users to select a certain Terraform version or OpenTofu.)