rust: undo wrong code for parallelism detection on NetBSD.
This mirrors the upstream pull request I just filed at

  https://github.com/rust-lang/rust/pull/116665

I have applied this to a 1.72.1 rust checkout, and have re-built
the bootstrap kits for our NetBSD ports with that change applied,
and this commit bumps the bootstrap requirement for 1.73.0 from
1.72.0 to use the newly generated 1.72.1 bootstrap kits for all
the NetBSD targets.

This gets us building again on NetBSD/evbarmv7hf 9.0, NetBSD/macppc
10.0_BETA, and most probably NetBSD/i386 9.3 as well, whereas with
the 1.72.0 bootstrap code, the "bootstrap" rust program would crash
with SEGV on those particular platforms.