newsboat: Import newsboat-2.33 as wip/newsboat233
Newsboat is a fork of Newsbeuter, an RSS/Atom feed reader for the
text console.  It runs on Linux, FreeBSD, Mac OS X and other
Unix-like operating systems.  Newsboat's great configurability and
vast number of features make it a perfect choice for people that
need a slick and fast feed reader that can be completely controlled
via keyboard.  It's the Mutt of RSS feed readers.