lima: Update to 0.18.0
Changes:
v0.18.0
-------
- `wsl2` driver (experimental)
- `limactl create --plain` to create "plain" instances (no mounts, port
  forwarding, containerd, etc.)
- Significant updates on the website (https://lima-vm.io/)

v0.17.2
-------
On macOS hosts, Lima now asks the user to sign the QEMU binary with the
`com.apple.security.hypervisor` entitlement if the binary is not
properly signed.

This signing is usually not needed on users' side.

v0.17.1
-------
This release fixes incompatibility with Go 1.21.

v0.17.0
-------
This release adds the `limactl create` command.
Now it is discouraged (not deprecated) to use `limactl start` for
creating new instances.

The `limactl create` command also takes several flags like `--cpus=1
--memory=2 --vm-type=vz --mount-writable=true` for modifying the
template.

Some of these flags are similar to the flags of `colima start`.

v0.16.0
-------
This release adds an experimental support for the `user-v2` network driver.
The `user-v2` network driver enables VM-to-VM networking without the
root privilege on the host.
This should be useful for simulating multi-node Kubernetes clusters on a laptop.

This release also fixes a relatively minor vulnerability CVE-2023-32684.
This vulnerability is very unlikely to be exploitable as long as you
are using the official templates of Lima.

v0.15.1
-------
This release adds support for QEMU 8.0.