ruby-rss: Import ruby-rss-0.3.0 as wip/ruby-rss
Really Simple Syndication (RSS) is a family of formats that describe
feeds, specially constructed XML documents that allow an interested
person to subscribe and receive updates from a particular web
service. This portion of the standard library provides tooling to read
and create these feeds.

The standard library supports RSS 0.91, 1.0, 2.0, and Atom, a related
format. Here are some links to the standards documents for these
formats:

* RSS
  * 0.9.1
  * 1.0
  * 2.0
* Atom