ruby-mustermann2: Import mustermann-2.0.2 as wip/mustermann2
Welcome to Mustermann.  Mustermann is your personal string matching expert.
As an expert in the field of strings and patterns, Mustermann keeps its
runtime dependencies to a minimum and is fully covered with specs and
documentation.

Given a string pattern, Mustermann will turn it into an object that behaves
like a regular expression and has comparable performance characteristics.