py-trio-websocket: import py-trio-websocket-0.11.1 as wip/py-trio-websocket
(py-trio-websocket is needed for the multimedia/streamlink update.)

This library implements both server and client aspects of the the WebSocket
protocol, striving for safety, correctness, and ergonomics. It is based on
the wsproto project, which is a Sans-IO state machine that implements the
majority of the WebSocket protocol, including framing, codecs, and events. This
library handles I/O using the Trio framework.