libplacebo: Import libplacebo-6.338.1 as wip/libplacebo
libplacebo is, in a nutshell, the core rendering algorithms and ideas
of mpv rewritten as an independent library. As of today, libplacebo
contains a large assortment of video processing shaders, focusing on
both quality and performance.

Fails at configure phase... we need to investigate regarding glad,
added some notes in TODO. (Also including `graphics/glad` bl3 fails in
the same way.)