libplacebo: Fetch <fast_float/fast_float.h> manually
Fetch <fast_float/fast_float.h> manually because it is expected to be
present (but we just mv it to the proper directory without needing
git).

Also bump USE_CXX_FEATURES to c++17 because it is needed for
`__has_include` (probably we could use just `has_include` instead of
c++17 but I think `has_include` is a subset of c++17 so just requiring
c++17 is probably fine and safer).