sops: Update to 3.8.1
Changes:
v3.8.1
------
In this release of SOPS, we have focused on landing a variety of bug
fixes to improve the overall user experience.

v3.8.0
------
We are extremely happy to introduce this new minor release of SOPS, now
a CNCF Sandbox project under the stewardship of a new group of
maintainers.

This release involved significant effort in rewriting and enhancing key
source implementations, and includes a number of bug fixes which people
had been patiently waiting on for a long period of time. In addition,
the release process has been built up from scratch, and now now
provides enhanced assurance by publishing SBOMs (Software Bill of
Materials), SLSA3 provenance attestations and a Cosign signed checksums
file as release artifacts.