mitmproxy: Import mitmproxy-8.1.1nb2 as wip/mitmproxy
This package contains tool for HTTP(S) sniffing.

mitmproxy is an interactive, SSL-capable man-in-the-middle proxy
for HTTP with a console interface.

mitmdump is the command-line version of mitmproxy. Think tcpdump
for HTTP.

mitmweb is a web-based interface for mitmproxy.

pathoc and pathod are perverse HTTP client and server applications
designed to let you craft almost any conceivable HTTP request,
including ones that creatively violate the standards.

Imported as is from net/mitmproxy. Intended to be used in order to update
net/mitmproxy.