loki: Update to 2.8.7
(This is the latest major 2.8.x release, an update to 2.9.x will be done in
separate commit(s))

Changes:
2.8.7
-----
This is release `v2.8.7` of Loki.

This release updates a few dependencies to bring in various security
patches.

2.8.6
-----
This is release `v2.8.6` of Loki.

Patches CVE-2023-39325 / CVE-2023-44487 by upgrading go to v1.20.10,
golang.org/x/net to v0.17.0 and grpc-go to v1.56.3

2.8.5
-----
This is release `v2.8.5` of Loki.

This release fixes a notable vulnerability to our published images.

* Updated all Docker base images to Alpine 3.18.3 to mitigate CVE-2022-48174 (#10573 )

2.8.4
-----
This is release `v2.8.4` of Loki.

This release fixes a few vulnerabilities in Loki and our published images.

* Fix CVE-2023-1255, CVE-2023-2650, CVE-2023-2975, CVE-2023-3446, and
  CVE-2023-3817
* Upgrade base alpine image used in loki image to 3.18.2

2.8.3
-----
This is release `v2.8.3` of Loki.

Loki:
* Upgrade go version to 1.20.6
* Querier: configurable chunk cache writeback queue bytes size
* Query-scheduler: fix query distribution in SSD mode
* Fix duplicate label values from ingester streams.

Promtail:
* Break on iterate journal failure
* Prevent logging errors on normal shutdown

2.8.2
-----
This is release `v2.8.2` of Loki.

- Upgrade to go 1.20.4 to address security vulnerabilities in the
  previous go version.
- Promtail: Add new `decompression` configuration to customize the
  decompressor behavior.

2.8.1
-----
We're excited to announce release `v2.8.1` of Loki. This release brings
a few bug fixes and security improvements.

* Fix bug that was dropping index if period is a zero value.
* Fix redis client to prevent it from incorrectly choosing cluster mode
  with local address.
* Update go to 1.20.3 to address security vulnerabilities in the
  previous go version.
* Update alpine image to 3.16.5 to address security vulnerabilities in
  the previous version.
* *Promtail*: Fix journald support in amd64 binary build.

For a full list of all changes please look at the [CHANGELOG](https://github.com/grafana/loki/blob/release-2.8.x/CHANGELOG.md).

2.8.0
-----
Grafana Labs is excited to announce the release of Loki 2.8. :tada:

- TSDB index no longer experimental: After extensive testing of the new
  TSDB index in our Grafana Cloud Logs offering, we can safely announce
  that it is no longer experimental and encourage all Loki deployments to
  use it!
- Query blocker: Queries can now be blocked in the Querier/Ruler via a
  per-tenant runtime configuration.
- New `backend` target: A new, third target was added to Loki's
  _scalable_ configuration, which is the default configuration used in
  the Loki helm chart. This allows Loki to be run as 3 targets (`read`,
  `write`, and `backend`) and makes the `read` target stateless and
  therefore able to be run as a Kubernetes deployment that can be scaled
  automatically.

As always, please read the upgrade guide at
<https://github.com/grafana/loki/blob/release-2.8.x/docs/sources/upgrading/_index.md>
before upgrading Loki.

2.7.5
-----
This is release `v2.7.5` of Loki, a patch release of the v2.7 minor
release.

*  Flush buffered logger on exit: this makes sure logs are printed if
   Loki crashes on startup.